refactor: decouple db client write from l2
