+++ .vscode/launch.json
@@ -35,11 +35,11 @@
     {
       "type": "node",
       "request": "launch",
-      "name": "l1/db write client",
+      "name": "db write client",
       "skipFiles": [
         "<node_internals>/**"
       ],
-      "program": "${workspaceFolder}/packages/client/src/write.ts",
+      "program": "${workspaceFolder}/packages/client/src/db.write.ts",
       "runtimeArgs": [
         "-r",
         "ts-node/register",
@@ -58,8 +58,6 @@
       "args": [
         "--resolver",
         "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512",
-        "-pk",
-        "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
       ]
     },
     {
@@ -69,7 +67,7 @@
       "skipFiles": [
         "<node_internals>/**"
       ],
-      "program": "${workspaceFolder}/packages/client/src/write.ts",
+      "program": "${workspaceFolder}/packages/client/src/l2.write.ts",
       "runtimeArgs": [
         "-r",
         "ts-node/register",
@@ -88,8 +86,6 @@
       "args": [
         "--resolver",
         "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512",
-        "-pk",
-        "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80",
         "--providerl2",
         "http://127.0.0.1:8545"
       ]
